requirements.txt
streamlit
Basic Knowledge Layer
faiss-cpu
numpy
sentence-transformers
requests
